Moved the file into IndivisualServicesFolder
